Fyp 117 investigate reward function (#64)
* fix typo in readme

* add friction to env

* increase traj len in sample model traj

* try new hanging reward

* add turns in state space

* add converted demo

* add converted state training

* add fully trained env

* remove checkpoints

* style fixes

+++ 04_Repository/Gym/Wrappers/hovering_wrapper.py
@@ -23,7 +23,7 @@ class HoveringWrapper(gym.Wrapper):
         obs, info = self.env.reset(seed, options, degrees, position)
         self.episode_count = 0
         # Return the augmented observation
-        augmented_obs = np.append(obs, self.get_hover_ratio())
+        augmented_obs = np.append(obs, self.get_hover_ratio()).astype(np.float32)
         return augmented_obs, info
 
     def get_hover_ratio(self):
@@ -44,5 +44,5 @@ class HoveringWrapper(gym.Wrapper):
         # print(self.hover_count, np.linalg.norm(action))
         # Augment the observation with the hover count
 
-        augmented_obs = np.append(obs, self.get_hover_ratio())
+        augmented_obs = np.append(obs, self.get_hover_ratio()).astype(np.float32)
         return augmented_obs, reward, done, truncated, info
